Up Nilai

+++ app/Http/Controllers/BE/Process/NilaiController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers\BE\Process;
 
 use App\Http\Controllers\BE\BaseController;
 use App\Services\BaseService;
+use App\Services\DB\MstBobotService;
 use App\Services\DB\MstEventService;
+use App\Services\DB\MstKriteriaService;
 use App\Services\DB\MstPesertaService;
 use App\Services\DB\NilaiService;
 use App\Services\DB\RegisterService;
@@ -51,27 +53,21 @@ class NilaiController extends BaseController
     public function DataNilai(Request $request)
     {
         try {
-            $data = [];
-            if (!empty($request->kd_event)) {
-                // Data Nilai
-                $dtnilai = NilaiService::Data("", false);
-                $dtnilai = $dtnilai->select("kd_event", "kd_peserta");
-                $dtnilai = $dtnilai->selectRaw("SUM(nilai)/COUNT(nilai) AS rata");
-                $dtnilai = $dtnilai->groupBy("kd_event", "kd_peserta");
-
-                $data = RegisterService::Data($request->kd_event);
-                $data = $data->leftJoinSub($dtnilai, "dn", function ($q) {
-                    $q->on("dn.kd_event", "=", "registerevent.kd_event");
-                    $q->on("dn.kd_peserta", "=", "registerevent.kd_peserta");
-                });
-                $data = $data->select("registerevent.kd_event", "registerevent.kd_peserta", "p.nm_peserta");
-                $data = $data->selectRaw("IFNULL(dn.rata, 0) AS rata, CASE WHEN dn.kd_peserta IS NOT NULL THEN 1 ELSE 0 END AS setup");
-                $data = $data->orderBy("p.nm_peserta")->get();
+            $validation = Validator::make($request->all(), ["kd_event" => "required", "kd_peserta" => "required"]);
+            if ($validation->fails()) {
+                $this->error = $validation->errors();
+                throw new \Exception(BaseService::MessageCheckData(), 400);
             }
+            $data = MstBobotService::Data($request->kd_event, true);
+            $data = NilaiService::Join($data, "mstbobot.kd_event", "mstbobot.kd_kriteria", $request->kd_peserta, "nd", "leftJoin", "v2");
+            $data = $data->select("mstbobot.kd_event", "mstbobot.kd_kriteria", "mk.nm_kriteria");
+            $data = $data->selectRaw("IFNULL(nd.nilai, 0) AS nilai");
+            $data = $data->orderBy("mk.nm_kriteria")->get();
             $this->respon = BaseService::ResponseSuccess(BaseService::MsgSuccess($this->pns, 1), $data);
         } catch (\Throwable $th) {
             $this->respon = BaseService::ResponseError($th->getMessage(), $this->error, $th->getCode());
         }
+        return $this->SendResponse();
     }
 
     public function Save(Request $request)
@@ -79,35 +75,48 @@ class NilaiController extends BaseController
         try {
             DB::beginTransaction();
             $validation = Validator::make($request->all(), [
-                "action" => "required|in:{$this->option_action}",
                 "kd_event" => "required",
                 "kd_peserta" => "required",
-                "tgl_register" => "required|date",
-                "no_event" => "required",
+                "dtnilai" => "required|array|min:1",
             ]);
             if ($validation->fails()) {
                 $this->error = $validation->errors();
                 throw new \Exception(BaseService::MessageCheckData(), 400);
             }
 
-            // Check Data event
+            // Check Data Event
             $dtevent = MstEventService::Detail($request->kd_event);
             // Check Data Peserta
             $dtpeserta = MstPesertaService::Detail($request->kd_peserta);
+            // Check Data Register
+            $regis = RegisterService::Detail($request->kd_event, $request->kd_peserta);
+
+            // Delete Data Detail Nilai
+            NilaiService::Data($request->kd_event, false)->where("nilaidetail.kd_peserta", $request->kd_peserta)->delete();
 
             // Save Or Update
-            $data = RegisterService::Detail($request->kd_event, $request->kd_peserta, $request->action);
-            if ($request->action == "Add") {
-                $data = RegisterService::new ();
+            $timeupdate = Carbon::now();
+            $totalNilai = 0;
+            $countNilai = 0;
+            foreach ($request->dtnilai as $val) {
+                // Check Kriteria
+                $dtnilai = MstKriteriaService::Detail($val["kd_kriteria"]);
+                // Check Kriteria in Bobot
+                $dtbobot = MstBobotService::Detail($request->kd_event, $val["kd_kriteria"]);
+
+                $data = NilaiService::new ();
                 $data->kd_event = $request->kd_event;
                 $data->kd_peserta = $request->kd_peserta;
-                $data->created_at = Carbon::now();
-            }
-            $data->no_event = $request->no_event;
-            $data->tgl_register = $request->tgl_register;
-            $data->updated_at = Carbon::now();
-            $data->save();
+                $data->kd_kriteria = $val["kd_kriteria"];
+                $data->nilai = $val["nilai"];
+                $data->created_at = $timeupdate;
+                $data->updated_at = $timeupdate;
+                $data->save();
 
+                $totalNilai += $val["nilai"];
+                $countNilai++;
+            }
+            $data = NilaiService::Data($request->kd_event, false)->where("nilaidetail.kd_peserta", $request->kd_peserta)->get();
             DB::commit();
             $this->respon = BaseService::ResponseSuccess(BaseService::MsgSuccess($this->pns, 2), $data);
         } catch (\Throwable $th) {
@@ -127,35 +136,19 @@ class NilaiController extends BaseController
                 throw new \Exception(BaseService::MessageCheckData(), 400);
             }
 
-            // Delete
-            $data = RegisterService::Detail($request->kd_event, $request->kd_peserta);
-            $data->delete();
+            // Check data
+            $dt = NilaiService::Data($request->kd_event, false)->where("nilaidetail.kd_peserta", $request->kd_peserta)->count();
+            if ($dt > 0) {
+                $data = NilaiService::Data($request->kd_event, false)->where("nilaidetail.kd_peserta", $request->kd_peserta);
+                $data->delete();
+            } else {
+                throw new \Exception(BaseService::MessageNotFound(), 400);
+            }
 
             DB::commit();
             $this->respon = BaseService::ResponseSuccess(BaseService::MsgSuccess($this->pns, 3), $data);
         } catch (\Throwable $th) {
             DB::rollBack();
-            $this->respon = BaseService::ResponseError($th->getMessage(), $this->error, $th->getCode());
-        }
-        return $this->SendResponse();
-    }
-
-    public function PesertaReady(Request $request)
-    {
-        try {
-            $validation = Validator::make($request->all(), ["kd_event" => "required"]);
-            if ($validation->fails()) {
-                $this->error = $validation->errors();
-                throw new \Exception(BaseService::MessageCheckData(), 400);
-            }
-            $data = MstPesertaService::Data();
-            $data = RegisterService::Join($data, $request->kd_event, "mstpeserta.kd_peserta", "re", "leftJoin", "v2");
-            $data = $data->whereNull("re.kd_peserta");
-            $data = $data->select("mstpeserta.kd_peserta", "mstpeserta.nm_peserta");
-            // dd($data->toSql());
-            $data = $data->orderBy("mstpeserta.nm_peserta")->get();
-            $this->respon = BaseService::ResponseSuccess(BaseService::MsgSuccess("Pserta Ready", 1), $data);
-        } catch (\Throwable $th) {
             $this->respon = BaseService::ResponseError($th->getMessage(), $this->error, $th->getCode());
         }
         return $this->SendResponse();
